feat : added theme toggle=>refactoring styles
