+++ package.json
@@ -25,7 +25,7 @@
   "lint-staged": {
     "src/**/*.{js,json,css,scss,md}": [
       "prettier --write",
-      "eslint --max-warnings=2"
+      "eslint --max-warnings=5"
     ]
   },
   "keywords": [],
