+++ src/components/DisplayComponent/styles.scss
@@ -6,15 +6,17 @@
   position: relative;
   justify-content: center;
   text-align: right;
+  background-color: var(--bg-color);
 
   &__result {
     width: 100%;
     height: 50%;
     padding-right: 15px;
+    color: var(--font-light);
   }
   &__calculation {
     font-size: 1rem;
-    color: #fff;
+    color: var(--font-light);
     padding-right: 15px;
   }
 }
