+++ src/styles.scss
@@ -1,59 +1,35 @@
-body {
-  height: 100vh;
-  width: 100vw;
-  background: #1939f6;
-  overflow: hidden;
+:root {
+  --bg-color: #171923;
+  --bg-light: #232535;
+  --font-color: #c5cddb;
+  --font-light: #ffffff;
 }
 
 * {
   box-sizing: border-box;
 }
 
-#app {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  font-family: 'Roboto', sans-serif;
-  font-weight: 300;
+body {
+  color: var(--font-light);
+  font-family: 'Montserrat', sans-serif;
+  background-color: var(--bg-color);
   font-size: 36px;
+  transition: all 500ms;
+  height: 100%;
 }
 
-.calculator {
-  opacity: 0;
-  animation: 1s bounceIn;
-  animation-fill-mode: forwards;
-  margin: 10px;
-  border-radius: 8px;
-  color: #ffffff;
-  box-shadow: 10px 10px 50px rgba(0, 0, 0, 0.5);
-  margin: auto;
-  height: 540px;
-  background: #191b1d;
+.lightMode {
+  --bg-color: #e8e6dc;
+  --bg-light: #dcdaca;
+  --font-color: #3d3d3d;
+  --font-light: #202020;
 }
 
-.buttons_container {
-  position: relative;
+#app {
   display: flex;
-
-  &__left {
-    display: grid;
-    grid: repeat(5, 67px) / repeat(3, 85px);
-  }
-  &__right {
-    display: grid;
-    grid: repeat(5, 67px) / repeat(1, 85px);
-  }
-}
-
-.btn_result {
-  position: absolute;
-  right: -12px;
-  bottom: -12px;
-  width: calc(43% * 1.25);
-  height: calc(100% * 0.25);
-  background: #1939f6;
-  border-bottom-right-radius: 8px;
-  border: 0.5px rgba(0, 0, 0, 0.0001);
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
 }
 
 @keyframes bounceIn {
